requirements.txt
Django==1.8.3
Pillow==2.9.0
boto==2.39.0
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==1.0.1
django-chartit==0.1
django-imagekit==3.3
django-mathfilters==0.3.0
django-mptt==0.7.4
django-polymorphic==0.7.1
django-polymorphic-tree==1.1
django-sql-explorer==0.7
django-storages==1.1.8
django-tag-parser==2.1
future==0.15.0
gunicorn==19.3.0
jsonfield==1.0.3
pilkit==1.1.13
psycopg2==2.6.1
pystache==0.5.4
six==1.10.0
sqlparse==0.1.11
static==1.1.1
django-mathfilters==0.3.0
static3==0.6.1
